requirements.txt
selenium
webdriver-manager
pyyaml
requests
beautifulsoup4
html5lib
pandas
spacy
